[cocoapods-bazel] Glob more dirs (#6)
expand_glob is overzealous when it tries to convert things from podspecs
into globs for bazel. If it doesn't match a few heuristics, it is
assumed to be a glob pattern of some sort that should be passed through
as-is *unless* it matches a few file extensions for directories. In that
case, '/**/*' is added to the end. This lead to things like
`s.exclude_files = "ios/Fabric"` in `react-native-safe-area-context`'s
podspec not getting converted to `exclude = ["ios/Fabric/**"]` in the
bazel file.

The main change we make is for the excludes on srcs, we expand
directories, and if there is no file extension on the pattern string,
and the pattern string resolves (literally, no glob resolution) to a
directory on disk, we say that it should get '/**/*' appended.